backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
sqlmodel>=0.0.14
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
python-dotenv>=1.0.0

# Authentication
pyjwt[crypto]>=2.8.0
passlib[argon2]>=1.7.4
cryptography>=42.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
